+++ CHANGELOG.md
@@ -7,11 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
  - Planar Reflection Probe support roughness (gaussian convolution of captured probe)
+ - Screen Space Refraction projection model (Proxy raycasting, HiZ raymarching)
+ - Screen Space Refraction settings as volume component
 
 ### Changed
  - Depth and color pyramid are properly computed and sampled when the camera renders inside a viewport of a RTHandle.
  - Forced Planar Probe update modes to (Realtime, Every Update, Mirror Camera)
  - Removed Planar Probe mirror plane position and normal fields in inspector, always display mirror plane and normal gizmos
+ - Screen Space Refraction proxy model uses the proxy of the first environment light (Reflection probe/Planar probe) or the sky
 
 ## [0.1.6] - 2018-xx-yy